tests/wpt/web-platform-tests/IndexedDB/idbfactory_cmp.htm

<!DOCTYPE html>
<meta charset=utf-8>
<title>IDBFactory.cmp() - compared keys return correct value</title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=support.js></script>

<script>
    test(function() {
        var greater = window.indexedDB.cmp(2, 1);
        var equal = window.indexedDB.cmp(2, 2);
        var less = window.indexedDB.cmp(1, 2);

        assert_equals(greater, 1, "greater");
        assert_equals(equal, 0, "equal");
        assert_equals(less, -1, "less");
    }, "IDBFactory.cmp()");
</script>

<div id=log></div>
